requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.16.18
botocore==1.19.18
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-celery-beat==2.1.0
django-countries==6.1.3
django-debug-toolbar==3.1.1
django-decouple==2.1
django-storages==1.10.1
django-timezone-field==4.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
kombu==5.0.2
oauthlib==3.1.0
paystackapi==2.0.0
Pillow==7.2.0
prompt-toolkit==3.0.8
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
